# Vectorized decoding in parquet reader