--- a/lib/trino-parquet/src/main/java/io/trino/parquet/reader/flat/DictionaryDecoder.java
+++ b/lib/trino-parquet/src/main/java/io/trino/parquet/reader/flat/DictionaryDecoder.java
@@ -29,18 +29,20 @@ public final class DictionaryDecoder<T>
     private final ColumnAdapter<T> columnAdapter;
     private final int dictionarySize;
     private final boolean isNonNull;
+    private final boolean vectorizedDecodingEnabled;
     private final long retainedSizeInBytes;
 
     private ValueDecoder<int[]> dictionaryIdsReader;
     @Nullable
     private Block dictionaryBlock;
 
-    public DictionaryDecoder(T dictionary, ColumnAdapter<T> columnAdapter, int dictionarySize, boolean isNonNull)
+    public DictionaryDecoder(T dictionary, ColumnAdapter<T> columnAdapter, int dictionarySize, boolean isNonNull, boolean vectorizedDecodingEnabled)
     {
         this.columnAdapter = requireNonNull(columnAdapter, "columnAdapter is null");
         this.dictionary = requireNonNull(dictionary, "dictionary is null");
         this.dictionarySize = dictionarySize;
         this.isNonNull = isNonNull;
+        this.vectorizedDecodingEnabled = vectorizedDecodingEnabled;
         this.retainedSizeInBytes = columnAdapter.getSizeInBytes(dictionary);
     }
 
@@ -48,7 +50,7 @@ public final class DictionaryDecoder<T>
     public void init(SimpleSliceInputStream input)
     {
         int bitWidth = input.readByte();
-        dictionaryIdsReader = new RleBitPackingHybridDecoder(bitWidth);
+        dictionaryIdsReader = new RleBitPackingHybridDecoder(bitWidth, vectorizedDecodingEnabled);
         dictionaryIdsReader.init(input);
     }
 
@@ -105,7 +107,8 @@ public final class DictionaryDecoder<T>
             DictionaryPage dictionaryPage,
             ColumnAdapter<BufferType> columnAdapter,
             ValueDecoder<BufferType> plainValuesDecoder,
-            boolean isNonNull)
+            boolean isNonNull,
+            boolean vectorizedDecodingEnabled)
     {
         int size = dictionaryPage.getDictionarySize();
         // Extra value is added to the end of the dictionary for nullable columns because
@@ -113,6 +116,6 @@ public final class DictionaryDecoder<T>
         BufferType dictionary = columnAdapter.createBuffer(size + (isNonNull ? 0 : 1));
         plainValuesDecoder.init(new SimpleSliceInputStream(dictionaryPage.getSlice()));
         plainValuesDecoder.read(dictionary, 0, size);
-        return new DictionaryDecoder<>(dictionary, columnAdapter, size, isNonNull);
+        return new DictionaryDecoder<>(dictionary, columnAdapter, size, isNonNull, vectorizedDecodingEnabled);
     }
 }