… docs/src/main/sphinx/object-storage/file-formats.md
… lib/trino-parquet/pom.xml
… lib/trino-parquet/src/main/java/io/trino/parquet/reader/decoders/ShortDecimalFixedWidthByteArrayBatchDecoder.java
… lib/trino-parquet/src/main/java/io/trino/parquet/reader/decoders/ValueDecoders.java
… lib/trino-parquet/src/main/java/io/trino/parquet/reader/decoders/VectorIntBitUnpackers.java
… lib/trino-parquet/src/main/java/io/trino/parquet/reader/flat/BitPackingUtils.java
… lib/trino-parquet/src/main/java/io/trino/parquet/reader/flat/FlatDefinitionLevelDecoder.java
… lib/trino-parquet/src/main/java/io/trino/parquet/reader/flat/NullsDecoder.java
… lib/trino-parquet/src/main/java/io/trino/parquet/reader/flat/NullsDecoders.java
… lib/trino-parquet/src/main/java/io/trino/parquet/reader/flat/VectorBitPackingUtils.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/AbstractColumnReaderBenchmark.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/AbstractColumnReaderTest.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/TestColumnReaderFactory.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/TestFlatColumnReaderRowRanges.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/TestInt96Timestamp.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/TestNestedColumnReader.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/TestNestedColumnReaderRowRanges.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/decoders/BenchmarkRleBitPackingDecoder.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/decoders/TestBitUnpackers.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/decoders/TestBooleanValueDecoders.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/decoders/TestFixedWidthByteArrayValueDecoders.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/decoders/TestRleBitPackingDecoderBenchmark.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/flat/BenchmarkFlatDefinitionLevelDecoder.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/flat/TestBitPackingUtils.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/flat/TestFlatColumnReader.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/flat/TestNullsDecoder.java
… lib/trino-parquet/src/test/java/io/trino/parquet/reader/flat/TestNullsDecoderBenchmark.java
… lib/trino-parquet/src/test/java/org/apache/parquet/column/values/dictionary/TestDictionaryWriter.java
… plugin/trino-delta-lake/src/main/java/io/trino/plugin/deltalake/DeltaLakePageSourceProvider.java
… plugin/trino-delta-lake/src/main/java/io/trino/plugin/deltalake/DeltaLakeSessionProperties.java